refactor(emotes-service): extract infra components package
Separating component definitions from main keeps the
entrypoint focused on orchestration only.
