+++ apps/emotes-service/infra/components/stateful.go
@@ -1,4 +1,4 @@
-package main
+package components
 
 import (
 	"emotes-service/infra/stack"
